Agents/health_agent/requirements.txt

langchain==0.3.10
langchain-openai==0.2.11
langchain-community==0.3.10
langchain-core==0.3.22
faiss-cpu==1.9.0.post1
sentence-transformers==3.0.1
fastapi==0.114.0
uvicorn==0.30.6
pydantic==2.8.2
python-dotenv==1.0.1
scikit-learn==1.6.1
xgboost==2.1.1
pandas==2.2.3
langsmith==0.1.147
uuid==1.30